Refactor bots to be similar in principle to components. That is, Bots no longer subclass Plugin but a plugin is provided that can expose them.
The idea here is to provide more flexibility with what can call a bot. For example, components can now easily expose bots as a JID within their domain, as illustrated in the example component.

Note that the bots are now configured differently in the configuration file, so this commit breaks any existing configurations involving bots. See the Eliza module's POD for what it should now look like.

git-svn-id: http://code.sixapart.com/svn/djabberd/trunk@551 43dd9337-660f-0410-bd32-e7601923a1a1