patch from Alexandre Dulaunoy <adulau@gmail.com> Subject: DJabberd.orig/lib/DJabberd/Presence.pm - support RFC XMPP (8.2.7) - resource must be not present in the jid
Hi Brad,

We had a small issue with djabberd regarding the from of the jid as it
was containing the resource.

DJabberd.orig/lib/DJabberd/Presence.pm

my $xml = DJabberd::XMLElement->new("", "presence", { '{}type' => 'subscribed',

-                                                          '{}from' =>
$from->as_string,
+                                                          '{}from' =>
$from->as_bare_string,

and we just replace the from with a bare string. It works now well and
avoid multiple entries for the same jid when subscribing (Philippe
Teuwen and I had the issue with psi).

Feel free to add it,

Thanks a lot for your great free software,

Best regards,

adulau




git-svn-id: http://code.sixapart.com/svn/djabberd/trunk/DJabberd@744 43dd9337-660f-0410-bd32-e7601923a1a1