Working towards fixing the bug where multiple resources of the same user are not made aware of one another.
However, this is still doing something a little odd and I've not been able to track it down yet. This works well enough that Psi
does indeed show the list of the other resources, but the two-resources test is showing up some sort of quirk with the
presence broadcasts. I'm really just committing this so that I can tinker without fear of screwing up what I've done so far...



git-svn-id: http://code.sixapart.com/svn/djabberd/trunk@759 43dd9337-660f-0410-bd32-e7601923a1a1